% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble_meso}
\alias{assemble_meso}
\title{Assemble mesocosm data file from onsite and lab data}
\usage{
assemble_meso(project = "soilplant", tofile = FALSE)
}
\arguments{
\item{project}{character choice of "soilplant" or "soil"}

\item{tofile}{logical print results to file?}
}
\description{
Assemble mesocosm data file from onsite and lab data
}
\examples{
\dontrun{
dt <- assemble_meso(tofile = TRUE)
}
}

