% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{clean_lab}
\alias{clean_lab}
\title{Clean lab data}
\usage{
clean_lab(dbname = "pc_eddlab.db", tablename = "eddlab",
  begindate = "2014-09-07")
}
\arguments{
\item{dbname}{character file.path of edd database}

\item{tablename}{character name of internal database table}

\item{begindate}{character date in YYYY-MM-DD format specifying a lower limit on the date of measurements to return}
}
\value{
An updated SQLite database with seperate tables for "meso" and "field" studies. Data remains in long rather than wide format?
}
\description{
Clean lab data
}
\details{
mdl flags are removed from result field
}
\examples{
\dontrun{
clean_lab()
}
}

