% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\name{clean_lims}
\alias{clean_lims}
\title{Clean LIMS lab data files}
\usage{
clean_lims(sumpathlist, proj = "field", pwsw = "all")
}
\arguments{
\item{sumpathlist}{list of file.paths to lims raw data files}

\item{proj}{string options are field, fieldbw, fieldfw, and meso}

\item{pwsw}{string choice of all, sw, pw}
}
\description{
Clean LIMS lab data files
}
\details{
EBs are discarded, only meso data has FDs
}

