% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\name{clean_p}
\alias{clean_p}
\title{Clean phosphorus data}
\usage{
clean_p(ppath = file.path("Raw", "lab", "phosphorus"))
}
\arguments{
\item{ppath}{character file.path to raw phosphorus data}
}
\description{
averages FDs, strips EBs
}
\examples{
\dontrun{
phosphorus <- clean_p(ppath = file.path("Raw", "lab", "phosphorus"))
}
}

