% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{cleanlab}
\alias{cleanlab}
\title{Clean LIMS lab data files}
\usage{
cleanlab(sumpathlist, proj = "field", pwsw = "all")
}
\arguments{
\item{sumpathlist}{list of file.paths to lims raw data files}

\item{proj}{string options are field, fieldbw, fieldfw, and meso}

\item{pwsw}{string choice of all, sw, pw}
}
\description{
Clean LIMS lab data files
}
\details{
EBs are discarded, only meso data has FDs
}

