% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onsite.R
\name{get_fieldonsite}
\alias{get_fieldonsite}
\title{Get field onsite data}
\usage{
get_fieldonsite(onsitepath = file.path("Raw", "onsite"))
}
\arguments{
\item{onsitepath}{character file.path to folder containing raw onsite data files}
}
\description{
Get field onsite data
}
\examples{
\dontrun{
fieldonsite <- get_fieldonsite(onsitepath = file.path("Raw", "onsite"))
}
}

