% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{get_mesolab}
\alias{get_mesolab}
\title{Get mesocosm lab data}
\usage{
get_mesolab(project = "soilplant", eddpath = file.path("Raw", "lab", "EDD"),
  sulfpath = file.path("Raw", "lab"), dbname = "pc_eddlab.db")
}
\arguments{
\item{project}{character choice of "soil" or "soilplant"}

\item{eddpath}{character folder.path to folder containing edd data}

\item{sulfpath}{character folder.path to folder containing sulfide data}

\item{dbname}{character file.path of edd database}
}
\description{
Get mesocosm lab data
}
\examples{
\dontrun{
mesolab <- get_mesolab(eddpath = file.path("Raw", "lab", "EDD"), sulfpath = file.path("Raw", "lab"))
}
}

