% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onsite.R
\name{get_mesoonsite}
\alias{get_mesoonsite}
\title{Get mesocosm onsite data}
\usage{
get_mesoonsite(folderpath = file.path("Raw", "onsite"),
  experiment = "SoilPlant")
}
\arguments{
\item{folderpath}{character file.path to onsite raw data}

\item{experiment}{character choice of "soilplant" or "soil"}
}
\description{
Get mesocosm onsite data
}

