% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mesokey}
\alias{mesokey}
\title{Key mesocosm treatment labels to an input data.frame}
\usage{
mesokey(dt)
}
\arguments{
\item{dt}{data.frame}
}
\description{
Key mesocosm treatment labels to an input data.frame
}

