% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcplot.R
\name{mesotsplot}
\alias{mesotsplot}
\title{plot mesocosm data}
\usage{
mesotsplot(dt, params = names(dt)[c(11)], pwsw = "pw", tofile = FALSE,
  addlegend = TRUE)
}
\arguments{
\item{dt}{data.frame}

\item{params}{character vector of column names}

\item{pwsw}{character choice of "pw" or "sw"}

\item{tofile}{logical save plot to file?}

\item{addlegend}{logical add legend?}
}
\description{
Plot mesocosm data as a time-series
}
\examples{
\dontrun{
dt <- read.csv("inst/extdata/mesoall.csv")
mesotsplot(dt, params = names(dt)[c(6, 8, 11, 12, 13, 16, 19, 20, 21)], pwsw = "pw", tofile = FALSE)
}
}

