% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcplot.R
\name{scplot}
\alias{scplot}
\title{scatter plot of peat collapse data}
\usage{
scplot(dt, params = c("ph", "LPH", "salinity", "CL.mglCaCO3"),
  rangecor = c("LPH > 6"), bwfw = "all", pwsw = "pw")
}
\arguments{
\item{dt}{data.frame}

\item{params}{provide a list of variables to plot. This list is parsed two at a time for scplot. Otherwise can be provided as a subset of data.frame names}

\item{rangecor}{provide a list of logical arguments to trim data outliers}

\item{bwfw}{character choice of "bw" or "fw"}

\item{pwsw}{character choice of "pw" or "sw"}
}
\description{
Create a scatter plot of peat collapse data
}
\examples{
\dontrun{
dt <- read.csv(list.files("/home/jose/Documents/Science/Data/peatcollapse/",
pattern = "fieldall*")[1], stringsAsFactors = FALSE)
scplot(dt)
}
}

