% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcplot.R
\name{tsplot}
\alias{tsplot}
\title{Time-series plot}
\usage{
tsplot(dt, params = names(dt)[c(9, 12, 13, 14, 16, 17, 18, 19, 20, 22, 23)],
  bwfw = "fw", pwsw = "pw", inout = "in", inclegend = TRUE,
  tofile = FALSE)
}
\arguments{
\item{dt}{data.frame}

\item{params}{character vector of column names}

\item{bwfw}{character choice if "fw" or "bw"}

\item{pwsw}{character choice of "pw" or "sw"}

\item{inout}{character choice of "in" or "out"}

\item{inclegend}{logical include legend?}

\item{tofile}{logical save plot to disk?}
}
\description{
Time-series plot
}

