% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble_field}
\alias{assemble_field}
\title{Assemble field data file from onsite and lab data}
\usage{
assemble_field(tofile = TRUE)
}
\arguments{
\item{tofile}{logical print results to file?}
}
\description{
Assemble field data file from onsite and lab data
}
\examples{
\dontrun{
dt <- assemble_field(tofile = TRUE)
}
}

