% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble_meso}
\alias{assemble_meso}
\title{Assemble mesocosm data file from onsite and lab data}
\usage{
assemble_meso(experiment = "soilplant", onsitepath = file.path("Raw",
  "onsite"), sulfpath = file.path("Raw", "lab"), eddpath = file.path("Raw",
  "lab", "EDD"), tofile = FALSE)
}
\arguments{
\item{experiment}{character choice of "soilplant" or "soil"}

\item{onsitepath}{character folder.path to folder containing onsite data}

\item{sulfpath}{character folder.path to folder containing sulfide data}

\item{eddpath}{character folder.path to folder containing edd data}

\item{tofile}{logical print results to file?}
}
\description{
Assemble mesocosm data file from onsite and lab data
}
\examples{
\dontrun{
meso <- assemble_meso(eddpath = file.path("Raw", "lab", "EDD"), sulfpath = file.path("Raw", "lab"), tofile = FALSE)
}
}

