% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{clean_p}
\alias{clean_p}
\title{Clean phosphorus data}
\usage{
clean_p(sumpath = NA)
}
\arguments{
\item{sumpath}{character file.path to raw phosphorus data}
}
\description{
averages FDs, strips EBs
}
\examples{
\dontrun{
dt <- clean_p()
}
}

