% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{create_labdb}
\alias{create_labdb}
\title{Create sqlite database of peatcollapse lab data}
\usage{
create_labdb(eddpath = "inst/extdata/Raw/lab/EDD", dbname = "eddlab")
}
\arguments{
\item{eddpath}{character file.path to edd data folder}

\item{dbname}{character file name of output SQLite database}
}
\description{
Create sqlite database of peatcollapse lab data
}
\examples{
\dontrun{
create_labdb()
}
}

