% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{get_fieldlab}
\alias{get_fieldlab}
\title{Get lab results from Field Study}
\usage{
get_fieldlab(fieldonsite, addlims = TRUE, limspath = "inst/extdata/Raw/lab")
}
\arguments{
\item{fieldonsite}{data.frame output of get_fieldonsite}

\item{addlims}{logical integrate lims data with edd data?}

\item{limspath}{character folder.path to folder containing lims data}
}
\description{
Get lab results from Field Study
}
\examples{
\dontrun{
dt <- get_fieldlab(fieldonsite = fieldonsite)
}
}

