% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{get_mesolab}
\alias{get_mesolab}
\title{Get mesocosm lab data}
\usage{
get_mesolab(project = "soilplant", dbname = "pc_eddlab.db")
}
\arguments{
\item{project}{character choice of "soil" or "soilplant"}
}
\description{
Get mesocosm lab data
}
\examples{
\dontrun{
clab <- get_mesolab(project = "soilplant")
}
}

