% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onsite.R
\name{get_mesoonsite}
\alias{get_mesoonsite}
\title{Get mesocosm onsite data}
\usage{
get_mesoonsite(experiment = "SoilPlant", dtpath = NA)
}
\arguments{
\item{experiment}{character choice of "soilplant" or "soil"}

\item{dtpath}{character file.path to mesocosm onsite raw data}
}
\description{
Get mesocosm onsite data
}

